Refactor(FileService.js) update and remove scripts

+++ src/app/services/FileService.js
@@ -1,17 +1,19 @@
 const fs = require('fs');
 const path = require('path');
 const { promisify } = require('util');
-const { readFile } = require('fs');
+const { readdir, readFile } = require('fs');
 const LogService = require('../services/LogService');
 
 const readFilePromise = promisify(readFile);
+const readDirPromise = promisify(readdir);
 
 class FileService {
   async destroy() {
     const uploads = path.resolve(__dirname, '..', '..', '..', 'uploads');
-    fs.readdir(uploads, (err, files) => {
+    const files = await readDirPromise(uploads);
+    if (files.length) {
       files.map(file => fs.unlink(`${uploads}/${file}`, (err) => {}));
-    });
+    }
   }
 
   async handleFile(file) {
@@ -34,55 +36,33 @@ class FileService {
           err,
         });
       }
-
-      // fs.readFile(filePath, 'utf-8', (err, data) => {
-      //   resolve(data);
-      // });
     });
   }
 
   async handleGames(lines) {
     const games = [];
-    let gameId = 1;
+    let game_id = 1;
     lines.forEach((line) => {
-      const initGame = this.handleInitGame({ line, gameId });
+      const initGame = this.handleInitGame({ line, game_id });
       if (initGame) {
         games.push(initGame);
-        gameId++;
+        game_id++;
       }
-      const gameKills = this.handleKills({ line, games, gameId });
+      const gameKills = this.handleKills({ line, games, game_id });
       if (gameKills) {
-        games[gameId - 2] = gameKills;
+        games[game_id - 2] = gameKills;
       }
     });
     const uniquePlayersGames = this.handleDuplicatedPlayers(games);
+    uniquePlayersGames.forEach((game) => {
+      game.kills = JSON.stringify(game.kills);
+      game.players = game.players.join(',');
+    });
     const savedGames = await this.saveGames(uniquePlayersGames);
     return savedGames;
-    // if (line.toLowerCase().includes('initgame')) {
-    //   games.push({
-    //     gameId,
-    //     players: [],
-    //     total_kills: 0,
-    //     kills: {},
-    //   });
-    //   gameId++;
-    // } else if (line.toLowerCase().includes('kill')) {
-    //   const killInfo = line.split(': ')[2];
-    //   const [, person1, person2] = /(.+) killed (.+) by/.exec(killInfo);
-    //   const game = { ...games[gameId - 2] };
-    //   game.total_kills += 1;
-    //   game.players.push(person1, person2);
-    //   if (person1 === '<world>' || person1 === person2) {
-    //     game.kills[person2] = game.kills[person2] ? game.kills[person2] - 1 : person2Kills - 1;
-    //     games[gameId - 2] = game;
-    //     return;
-    //   }
-    //   game.kills[person1] = game.kills[person1] ? game.kills[person1] + 1 : person1Kills + 1;
-    //   games[gameId - 2] = game;
-    // }
   }
 
-  handleKills({ line, games, gameId }) {
+  handleKills({ line, games, game_id }) {
     if (!line.toLowerCase().includes('kill')) {
       return;
     }
@@ -90,7 +70,7 @@ class FileService {
     const person2Kills = 0;
     const killInfo = line.split(': ')[2];
     const [, person1, person2] = /(.+) killed (.+) by/.exec(killInfo);
-    const game = { ...games[gameId - 2] };
+    const game = { ...games[game_id - 2] };
     game.total_kills += 1;
     game.players.push(person1, person2);
     if (person1 === '<world>' || person1 === person2) {
@@ -101,12 +81,12 @@ class FileService {
     return game;
   }
 
-  handleInitGame({ line, gameId }) {
+  handleInitGame({ line, game_id }) {
     if (!line.toLowerCase().includes('initgame')) {
       return;
     }
     return {
-      gameId,
+      game_id,
       players: [],
       total_kills: 0,
       kills: {},
